fix: build OmniRigTransceiver on Windows
    OmniRigTransceiver.cpp:257: error: narrowing conversion of '-5' from
    'int' to 'long long unsigned int' [-Wnarrowing]

The switch compares a frequency reported by the rig against one just set, and
every case is a small delta — three of them negative. Whether the expression
is signed depends on the return type of the OmniRig binding dumpcpp generates,
and that differs between the type library WSJT-X points at with
OMNIRIG_TYPE_LIB and the COM class this tree resolves from the registry.
Unsigned, each negative case is a narrowing conversion and the build stops.

Cast both operands to qint64, which is what the comparison meant in the first
place.

This file compiles only on Windows, so no amount of local Linux building would
have found it — the third inherited Windows-only failure in a row, after the
.tls_common directive and the HRDMessage warning.

+++ Transceiver/OmniRigTransceiver.cpp
@@ -252,7 +252,13 @@ int OmniRigTransceiver::do_start ()
         {
           throw_qstring (tr ("OmniRig: don't know how to set rig frequency"));
         }
-      switch (rig_->GetRxFrequency () - test_frequency)
+      // Explicitly signed: the cases below are small negative deltas, and
+      // whether this expression is signed depends on the return type of the
+      // OmniRig binding dumpcpp generates — which differs between the type
+      // library WSJT-X points at and the COM class registered here. Unsigned,
+      // every negative case is a narrowing conversion and the build stops.
+      switch (static_cast<qint64> (rig_->GetRxFrequency ())
+              - static_cast<qint64> (test_frequency))
         {
         case -5: resolution = -1; break;  // 10Hz truncated
         case 5: resolution = 1; break;    // 10Hz rounded
